Rename bounds.h to box.h, Xcode and Autotools builds for box.h.
